[MTA] Add call function to call exported functions
The call function is (currently) the only function in the MTA's class prototype, it allows
you to easily call http exported functions by the resource's name and the function's name.
It also supports sending and returning data.

The function throws errors if the arguments are incorrect, and it uses a callback to return
any errors or data when requesting.